requirements.txt
aiosmtpd==1.4.4.post2
alembic==1.13.0
atpublic==4.0
attrs==23.1.0
Babel==2.14.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.4.2
docker==7.0.0
elastic-transport==8.11.0
elasticsearch==8.11.1
email-validator==2.1.0.post1
Flask==2.3.3
flask-babel==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
langdetect==1.0.9
Mako==1.3.0
MarkupSafe==2.1.3
packaging==23.2
pip-system-certs==4.0
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
rq==1.15.1
setuptools==69.0.3
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==2.3.8
wrapt==1.16.0
WTForms==3.1.1
